plugins.idp: extract base idp class for helpers
